requirements.txt
pymavlink>=2.4.40
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0
